add filename, filesize and mimetype handling to
the imagelist attribute and add the width and height of an image as well.
This means the filename, filesize, mimetype and width/height of an image
will be displayed after (async) uploading and may be modified when the
template of validation allows for it (as the imagelist attribute handles
the list of Image valueobjects as a value object).

The user provided original filename will be sanitized due to security
reasons before sending it back to the client side. The storage and
modification of this filename field doesn't change the honeybee behaviour
of using the location property with a randomized name for the actual
filename being used when downloading the (image) file. The filename was
added to make it available later on for searching or similar requirements.