add hidden input for boolean attribute checkbox
to always send a value from the web via POST when the checkbox should be
set or unset. Otherwise it wouldn't be possible to ever uncheck the
checkbox (that is, set the boolean attribute to the default FALSE) as the
unchecked input type=checkbox will not be sent by browsers according to
the spec.