configure user images to use different activities
for the thumbnail and preview images in the image widget of the images
attribute. This is necessary to have independant urls for the images of
the user and the rest of the system. The default converjon.thumbnail
activity can thus generate urls to e.g. AWS S3 file storage whileas user
images that are always stored locally (see the filesystems.xml in
Honeybee_SystemAccount module) can be delivered from the local filesystem.