handle GeoPoint empty payload specifically
The GeoPoint payload might be an array with ['lon'=>'0','lat'=>'0'] or
only one of the coordinates being 0. The default filterEmptyPayload method
removes array rows that are empty and as that php method treats '0' and 0
as empty values we need to take care of that as the actual
GeoPointAttribute might perfectly well accept the Null Island (0,0) value
or coordinates on the Prime Meridian or the Equator.