always add width/height information on upload
prior to this change the uploading of a file within the POST data would
always set the width/height to 0 as the values are not part of the
getHoneybeeProperties() array as width and height are not always available
for file handling attributes. Only Image file handling attributes have the
width/height properties. That's why another bit of code is needed in the
application's command validation to cater for those non-javascript enabled
upload-was-done-async-in-background cases where the file is part of the POST.