add connection status info to status page
As the connectors are now able to return status information this is used
to output an informative page. Output formats implemented for the moment
are console and json and to some degree the xml and html ones.

bin/cli honeybee.core.status
bin/cli honeybee.core.status -verbose (or -v)

//honeybee-core/status
//honeybee-core/status.json
//honeybee-core/status.xml