add: console/json/html/xml outputformat for Status
The Status action now asks for the connections' StatusReport and decides
whether the application is working or not depending on the returned info.

At the moment the connection infos are the only one taken into account
when deciding whether the application is working correctly. When no
failing connection status are returned, the app is deemed working.

The following output formats are supported:

Console:
bin/cli honeybee.core.status
bin/cli honeybee.core.status -v
bin/cli honeybee.core.status -verbose

HTML:
https://honeybee-cms.local/de/honeybee-core/status
https://honeybee-cms.local/de/honeybee-core/status?v=1
https://honeybee-cms.local/de/honeybee-core/status?verbose=true

XML:
https://honeybee-cms.local/de/honeybee-core/status.xml
https://honeybee-cms.local/de/honeybee-core/status.xml?v=1
https://honeybee-cms.local/de/honeybee-core/status.xml?verbose=true

JSON:
https://honeybee-cms.local/de/honeybee-core/status.json
https://honeybee-cms.local/de/honeybee-core/status.json?v=1
https://honeybee-cms.local/de/honeybee-core/status.json?verbose=true

Content negotiation for the "/honeybee-core/status" URL should work as
expected for the above output formats.

At the moment the output is generated in the SuccessView and does not use
renderers or renderSubject. This may be added later if practical.