status page is now under /status[.json|.xml][?v=1]
The URL of the status page is now stable. There are no longer locale
dependant versions of the URL (/de/honeybee-core/status etc.). Blocking
the access to the status page via webserver or load balancer should now
be easier as the monitoring doesn't need to know the supported locales
of the application anymore (when "app_status.is_secure=false").

The status page route is no longer defined in the Honeybee_Core module's
routing, but defined in application's routing. Projects can change the
module and action of the /status URL via system_action definitions in
their settings file.