FPF repopulates image data after failed validation
this was not working before as the first uploaded image had a form input
name with index [2] instead of [1] ([0] is the template and non-js
variant) and the gap ([1]) is missing after the view gets rendered on the
serverside before the FPF kicks in again. There's still the possibility
of gaps when images are uploaded and deleted on client side. The js widget
should handle this properly. This is not yet covered.