switch from yaml local configs to json format
this switch is necessary atm as our puppet setup generates local config
files that are valid yaml, but symfony yaml parsers doesn't think so.
Switching to libyaml and the pecl yaml extension is an option but would
need more work at the moment to get the appropriate puppet module and
binary repos…

We might actually suggest using pecl-yaml instead of symfony yaml parser
and use that when it is available?