Fix bug on multiple widget instances
When having multiple instances of a widget then the 'prefix' option would
be removed from the first instance and would not be available to the following
ones, falling back to the Widget prefix.
This is happening because the Require.js module is only being loaded once and
then being used for creating the actual widget instances.