modify email settings for user mail service and
add some german error translations as a start to improve things.

The text of the password reset email has been slightly changed as well and
the 20 minutes validity remark has been removed. The token_expire_date
though has not been removed yet from the reset password functionality and
the entity as of now (but should be in my opinion).