BC: default role administrator => full-privileged
refs #7

The restriction to a fixed set of roles in the User schema has been removed.

A simple AclRolesRenderer has been added to display the available system
roles on the modify view.

The theme preview page is now only accessible when logged in. A credential
has been defined that needs to be given to all not full-privileged roles.