use rel="noopener noreferrer" on target _blank As
This should prevent "window.opener" related issues when malicious links
find their way into the CMS and are clicked on by editors. With
"window.opener" set, a page can trigger a navigation in the opener
regardless of security origin. This commit doesn't help in current
safari or older browsers though as they don't support the rel attribute
value handling. Suggestion for CMS is to not use _blank if one wants to
make sure to be not affected.