add asset-list handling and improve upload errors
the rendering of generic uploaded files is more or less a copy of the
existing image-list rendering and widget without moving the assets
up/down and without the thumbnail image as that's not necessary right now.

File upload error messages have been improved and are now translatable.

For the moment the image-list widget now shows an alert message when
uploading fails due to server side issues or when the uploaded file does
not match the constraints defined in the upload validation xml file.