default to rendering full reference projections in
haljson output format. The view template usage is optional and can be used
to restrict the amount of rendered fields per entity in the reference list
attribute.

example: view template name is "foo.bar.baz" as the Baz entity from the
Foo_Bar module is being rendered in haljson. When only that view template
exists the full toArray() representation of reference list attributes of
the Baz entity will be output. When a view template for e.g. the Blah
entity type on the reference field "Baz::blub" is being added (like
"foo.bar.baz.blub.blah") then that entity of the reference is rendered as
specified by the view template while other entity types on that blub
reference attribute will still be rendered via toArray().