add workflow state specific activity label & desc
ription – it's now possible to translate the transition names of the
resource's workflow in a state (node) specific way. It falls back to the
previously used "activity.label". E.g.:

In a workflow where the resource is in state "published" and might
"promote" to archive and "demote", "delete" etc. it's possible to
define for translation domain "[vendor.package.type].activity":

- "published.promote.label"
- "promote.label"
- "published.promote.description"
- "promote.description"

That is, if there are multipe states where one can "promote" the resource
it's now possible to be explicit when translating a specific transition.
The fallback allows for situations where always the same label is okay.