fix: cache scope for every twig renderer instances
the default behaviour is only necessary when StringLoader and FileLoader
are used in the same process. As that only happens in the skeleton
generator the default behaviour is changed with this commit. The
previously used precaution might be better for correctness and hard to
debug template reusage within the same process, but leads to new twig
cache entries for each new process, a.k.a. http request and leads to
practically a disabled cache that fills your hard disk. :x