add: return to previously seen collection view url
with all filter/sort/search parameters set. The last seen collection view
url is stored in the user session and is then used to redirect after post
upon successful create or task/proceed actions. The "back_to_collection"
activity that exists in the skeleton is changed to return to that stored
url as well. That way the primary activity "back_to_collection" activity
becomes actually usefull...