allow readonly actions on final workflow states
prior to this change not only transitions were ignored on final workflow
states, but also all defined read_only_actions. As e.g. viewing the latest
data or the history of items might be helpful those actions are now
allowed in the WorkflowActivityService. This leads to links being rendered
on the collection views of entities w/ workflows that have readonly
activities defined on final workflow states.