bump to honeybee 0.2.12 for trellis BaseObject
the dependencies are bumped and twig is now 2.4.4 in the composer lock
file. Minimum honeybee/honeybee has been set to 0.2.12 as that version
brings trellis BaseObject instead of Object and thus should be used from
now on. The classes that use that base Object in this project have been
updated to use BaseObject.