chg behaviour of /status and /status/health checks
The StatusAction always checks connectors while the HealthAction only
does this when verbose parameter is true on cli or web. This change leads
to the following cases:

- /status => FAILING if at least one connector fails, UNKNOWN/WORKING
  otherwise (depending on all connector (fake) status checks)
- /status/health => HTTP-200 w/ "UNKNOWN" (as no checks are done)
- /status/health?v=1
    - => HTTP-500 w/ "FAILING" (at least one connector status check fails)
    - => HTTP-200 w/ "WORKING" (all connectors return WORKING)
    - => HTTP-200 w/ "UNKNOWN" (no FAILING, but some UNKNOWN connectors checks)

This means, that you need to configure all status checks to get WORKING or
use the "fake_status" parameter to WORKING for connectors that you don't
care about exactly.

The commandline versions behave similar:
- bin/cli health => exitcode -1 on at least one FAILING connector
- bin/cli health => exitcode 0 when status is not FAILING