revert 621463ad8a1e404368e51f555070dc2a90403592 &
bump squire-js and dompurify javascript libs. Change the richtext widget:

The sync config change breaks the functionality of adding simple line
breaks in a way that is not consistent or even understandable by editors
as the "div class=hb-paragraph" is being stripped and breaks by editors
are gone after saving etc.