Add selectize as NPM dependency
Point to a custom version from the Honeybee organization.
That version merges all the last fixes (up to the current date)
and includes some patches that atm are not yet merged in the
original Selectize repository.

TODO: Restore the original repository as dependency, in case our patches
get merged there:

    - remove custom repository from package.json and run

        npm install --save selectize